+++ install.py
@@ -3,7 +3,6 @@ import os
 
 # os.mkdir('log')
 
-
 # os.mkdir('src')
 # os.mkdir('src/preprocessing')
 # os.mkdir('src/preprocessing/library')
@@ -11,11 +10,10 @@ import os
 # os.mkdir('src/visualize')
 # os.mkdir('src/visualize/library')
 
-
-# os.mkdir('view')
-# os.mkdir('view/boxplot') 
-# os.mkdir('view/heatmap')
-# os.mkdir('view/plot') 
-# os.mkdir('view/scatter')
-# os.mkdir('view/heatmap-scatter') 
-# os.mkdir('view/violinplot') 
+os.mkdir('view')
+os.mkdir('view/boxplot') 
+os.mkdir('view/heatmap')
+os.mkdir('view/heatmap-scatter') 
+os.mkdir('view/line') 
+os.mkdir('view/scatter')
+os.mkdir('view/violinplot')