+++ src/visualize/scatter.py
@@ -1,23 +1,37 @@
+import logging
 import pandas as pd
 import matplotlib.pyplot as plt
 
-
-PATH = 'database/preprocessing.csv'
+from library.Const import Const
 
 
 if __name__ == "__main__" :
-    HEADER = ('Vazão', 'Temperatura', 'pH', 'Condutividade', 'OD',
-              'HCO3', 'MPS','Clorofila', 'COD',
-              'NO', 'NO.1', 'NH', 'NI', 'N', 'NT', 'NOD',
-              'NP', 'PO4', 'PTD', 'PT', 'PO', 'PP', 'Silica', 
-              'Cl', 'SO4', 'Na', 'Ca', 'K', 'Mg')
-
-    len_header = len(HEADER)
-    df = pd.read_csv(PATH)
+    logging.basicConfig(filename=('log/view-scatter.txt'),
+                    level=logging.DEBUG, 
+                    format=' %(asctime)s - %(levelname)s - %(message)s')
+    logging.disable()
+    logging.debug('Start of program')
+    
+    const = Const()
+    df = pd.read_csv(const.path)
+    del df['YYYY']
+    del df['class']
+    
+    header = []
+    for i in df.columns: 
+        header.append(i)
+    
+    len_header = len(header)
+    
+    logging.debug('PATH: ' + const.path)
+    logging.debug('header.len: ' + str(len_header))
+    logging.debug('header: ' + str(header))
     
     for i in range(len_header):
         for j in range(len_header):
-            df.plot.scatter(x = HEADER[i], y = HEADER[j], c='DarkBlue')
-            plt.savefig('view/scatter/' + HEADER[i] + '-vs-' + HEADER[j] +'.pdf', dpi=300) 
+            df.plot.scatter(x = header[i], y = header[j], c='DarkBlue')
+            plt.savefig('view/scatter/' + header[i] + '-vs-' + header[j] +'.pdf', dpi=300) 
             plt.cla()   # Clear axis
             plt.clf()   # Clear figure
+            
+    logging.debug('End of program')